Code cleanup and screen off timer
A few updates:
Cleaned up the calculator.h code

Added a feature that turns off the screen backlight after 30 seconds (this can be adjusted in the .ino file)

Fixed a bug in the right/left rotate function, so that should work now.

Added a 24-bit version of word flip and byte flip that rotates bytes. This allows you to swap color channels in 24-bit color mode.